==> 3-process_shapley.ipynb <==
import pandas as pd 
import numpy as np


def process_shapley(modAlgo, confNum):
    shapley_train = []
    shapley_test = []
    shapley_index = []
    for conf in range(0, confNum):
        print(conf)
        for dim in [5, 30]:
            for budget in [50*dim, 100*dim, 300*dim, 500*dim, 1000*dim, 1500*dim]: 
                for fold in range(1, 6):
                    df_tr = pd.read_csv(f"./results/{modAlgo}/shapley/shapley_budget_{budget}_conf_{conf}_{dim}D_fold_{fold}_train.csv")
                    df_te = pd.read_csv(f"./results/{modAlgo}/shapley/shapley_budget_{budget}_conf_{conf}_{dim}D_fold_{fold}_test.csv")
                    if fold != 1:
                        df_train = pd.concat([df_train, df_tr])
                        df_test = pd.concat([df_test, df_te])
                    else: 
                        df_train = df_tr
                        df_test = df_te
                df_train.index = df_train.iloc[:,0]
                df_train.drop(["Unnamed: 0"], inplace=True, axis=1)
                df_test.index = df_test.iloc[:,0]
                df_test.drop(["Unnamed: 0"], inplace=True, axis=1)
                data_train = []
                data_test = []
                for fid in range(1, 25):
                    for iid in range(1, 6):
                        df_train_sub = df_train[df_train.index== f'{fid}_{iid}']
                        df_test_sub = df_test[df_test.index== f'{fid}_{iid}']
                        data_train.append(list(np.mean(df_train_sub, axis=0)))
                        data_test.append(list(np.mean(df_test_sub, axis=0)))

                shapley_train.append(list(np.mean(data_train, axis=0)))
                shapley_test.append(list(np.mean(data_test, axis=0)))
                shapley_index.append(f'{conf}_{budget}_{dim}')
    df_shapley_train = pd.DataFrame(shapley_train, index=shapley_index)
    df_shapley_test=pd.DataFrame(shapley_test, index=shapley_index)
    df_shapley_train.to_csv(f"./results/{modAlgo}/shapley_merged/shapley_train.csv")
    df_shapley_test.to_csv(f"./results/{modAlgo}/shapley_merged/shapley_test.csv")

# process_shapley('modCMA', 324)
process_shapley('modDE', 576)
