==> 6-create_conf_per2vec.ipynb <==
import pandas as pd
import numpy as np


def create_per2vec_modCMA(numConf):
    dims = [5, 30]
    data = []
    data_index = []
    for conf in range(0, numConf):
        print(conf)
        for dim in dims: 
            conf_df = pd.read_csv(f'./data/raw_data/Performance_tables_modcma/conf_{conf}_{dim}D.csv', index_col=0)
            budgets = [50*dim, 100*dim, 300*dim, 500*dim, 1000*dim, 1500*dim] 
            for budget in budgets:
                conf_df_budget = conf_df[(conf_df['runtime']==budget)]
                arr_conf = []
                for FID in range(1, 25):
                    conf_df_budget_FID = conf_df_budget[(conf_df_budget['funcId']==FID)]
                    arr_FID = []
                    for IID in range(1, 6):
                        conf_df_budget_FID_IID = conf_df_budget_FID[(conf_df_budget_FID['ID']==IID)].iloc[0].values[4:14].tolist()
                        median_runs = np.median(conf_df_budget_FID_IID)
                        arr_FID.append(median_runs)
                    mean_instances = np.mean(arr_FID)
                    arr_conf.append(mean_instances)
                data.append(arr_conf)
                data_index.append(f'{conf}_{budget}_{dim}')

    df_conf_to_vec = pd.DataFrame(data=data, index = data_index, columns=['v_1', 'v_2', 'v_3', 'v_4', 'v_5', 'v_6', 'v_7', 'v_8', 'v_9', 'v_10', 'v_11', 'v_12', 'v_13', 'v_14', 'v_15', 'v_16', 'v_17', 'v_18', 'v_19', 'v_20', 'v_21', 'v_22', 'v_23', 'v_24'])
    df_conf_to_vec.to_csv('./data/processed_data/conf_perf2vec.csv')

def create_per2vec(modAlgo, numConf, log_mode):
    dims = [5, 30]
    data = []
    data_index = []
    for conf in range(0, numConf):
        print(conf)
        for dim in dims: 
            budgets = [50*dim, 100*dim, 300*dim, 500*dim, 1000*dim, 1500*dim] 
            for budget in budgets:
                conf_df_budget = pd.read_csv(f'./data/performance_data/{modAlgo}/{log_mode}/budget_{budget}_conf_{conf}_{dim}D.csv', index_col=0)
                arr_conf = []
                for FID in range(1, 25):
                    conf_df_budget_FID = conf_df_budget[conf_df_budget.index.str.startswith(str(FID)+"_")]
                    mean_instances = np.mean(conf_df_budget_FID['target'])
                    arr_conf.append(mean_instances)
                data.append(arr_conf)
                data_index.append(f'{conf}_{budget}_{dim}')

    df_conf_to_vec = pd.DataFrame(data=data, index = data_index, columns=['v_1', 'v_2', 'v_3', 'v_4', 'v_5', 'v_6', 'v_7', 'v_8', 'v_9', 'v_10', 'v_11', 'v_12', 'v_13', 'v_14', 'v_15', 'v_16', 'v_17', 'v_18', 'v_19', 'v_20', 'v_21', 'v_22', 'v_23', 'v_24'])
    df_conf_to_vec.to_csv(f'./data/classification_data/{modAlgo}_conf_perf2vec_{log_mode}.csv')

# create_per2vec('modCMA', 324, 'log')
# create_per2vec('modCMA', 324, 'unlog')

create_per2vec('modDE', 576, 'log')
create_per2vec('modDE', 576, 'unlog')
