==> 8-UMAP_cluster.ipynb <==
import pandas as pd
import pickle 
from umap import UMAP
import matplotlib.pyplot as plt
import plotly.express as px
import numpy as np
from denseclus import DenseClus

def get_modular_conf(modAlgo, module):
    setup = pd.read_csv(f'./data/raw_data/{modAlgo}_conf_grid.csv', index_col=0)[module]
    return list(setup)

def UMAPS(dim, budget, modAlgo):
    seed = 42
    df = pd.read_csv(f"./data/classification_data/{modAlgo}_conf_perf2vec_unlog.csv", index_col=0)
    print(df)
    variable = f'.*_{budget}_{dim}'           
    df = df[df.index.str.contains(fr'\b{variable}\b', regex=True)]
    if modAlgo == 'modCMA':
        module = 'elitist'
    else: 
        module = 'lpsr'
    confs = get_modular_conf(modAlgo,module)
    sv_2d = UMAP(
        n_components=3,
        n_neighbors=50,
        min_dist=0.0001,
        spread= 1,
        random_state=seed,
    ).fit_transform(df)
    sv_2d_df = pd.DataFrame(sv_2d, columns=["A", "B", "C"])
    sv_2d_df[module] = confs
    fig = px.scatter_3d(sv_2d_df, x="A", y="B", z="C", color = module)
    fig.update_traces(marker=dict(size=4,
                              line=dict(width=1,
                                        color='DarkSlateGrey')),
                  selector=dict(mode='markers'))
    fig.update_layout(
        width=500,
        height= 300,
        margin=dict(
        l=0,
        r=0,
        b=0,
        t=0,
        ),
        font = dict(
            size=10
        )
    )
    fig.update_scenes(
        xaxis_range=[-5, 15],
        yaxis_range=[0, 14],
        zaxis_range=[0, 13]
    )

    fig.show()

UMAPS(5, 1500, 'modCMA')
# UMAPS(5, 1500, 'modDE')
