==> 9-evaluate_clasification_algos.ipynb <==
import pandas as pd
import numpy as np

print("Prediction of modular configuration using the Shapley meta-representations.")
print("F1 scores aggregated across all modules, dims, budgets.")


print("====SHAP====")
print("\nModCMA")
f1_RF = pd.read_csv("./results/modCMA/classification_s2v_RF.csv", index_col=0)['f1']
print("RF",round(np.mean(f1_RF),3))
f1_RF = pd.read_csv("./results/modCMA/classification_s2v_RF_MLC.csv", index_col=0)['f1']
print( "RF-MLC",round(np.mean(f1_RF),3))
f1_TabPFN = pd.read_csv("./results/modCMA/classification_s2v_TabPFN.csv", index_col=0)['f1']
print("modCMA", "TabPFN",round(np.mean(f1_TabPFN),3))


print("\nmodDE")
f1_RF = pd.read_csv("./results/modDE/classification_s2v_RF.csv", index_col=0)['f1']
print("RF",round(np.mean(f1_RF),3))
f1_RF = pd.read_csv("./results/modDE/classification_s2v_RF_MLC.csv", index_col=0)['f1']
print( "RF-MLC",round(np.mean(f1_RF),3))
f1_TabPFN = pd.read_csv("./results/modDE/classification_s2v_TabPFN.csv", index_col=0)['f1']
print("TabPFN",round(np.mean(f1_TabPFN),3))



print("====performance====")
print("\nModCMA")
f1_RF = pd.read_csv("./results/modCMA/classification_p2v_RF.csv", index_col=0)['f1']
print("RF",round(np.mean(f1_RF),3))
f1_RF = pd.read_csv("./results/modCMA/classification_p2v_RF_MLC.csv", index_col=0)['f1']
print( "RF-MLC", round(np.mean(f1_RF),3))
f1_TabPFN = pd.read_csv("./results/modCMA/classification_p2v_TabPFN.csv", index_col=0)['f1']
print("modCMA", "TabPFN", round(np.mean(f1_TabPFN),3))


print("\nmodDE")
f1_RF = pd.read_csv("./results/modDE/classification_p2v_RF.csv", index_col=0)['f1']
print("RF", round(np.mean(f1_RF),3))
f1_RF = pd.read_csv("./results/modDE/classification_p2v_RF_MLC.csv", index_col=0)['f1']
print( "RF-MLC", round(np.mean(f1_RF),3))
f1_TabPFN = pd.read_csv("./results/modDE/classification_p2v_TabPFN.csv", index_col=0)['f1']
print("TabPFN", round(np.mean(f1_TabPFN),3))
